requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
narwhals==2.5.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
preshed==3.0.10
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.49.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
